Create rule form component
